fix: long -> int

+++ src/codegen/types.rs
@@ -4,8 +4,13 @@ use crate::ir::Type;
 pub fn type_to_c(_type: &Type) -> anyhow::Result<String> {
     match _type {
         Type::Char => Ok("char".to_string()),
-        Type::Number => Ok("long".to_string()),
+
+        // this is bad: `number` in js is 64 bits, `int` in C is 32. realistically this should be a long
+        Type::Number => Ok("int".to_string()),
+
+        // TODO: make it so that if this type ever occurrs in the code, we #include <stdbool.h> at the start of the program
         Type::Boolean => Ok("bool".to_string()),
+
         Type::Void => Ok("void".to_string()),
 
         Type::Pointer(inner) => {
